refactor: move ScannerUI outside of core.ui
This fixes import loop errors when trying to import Commit.
